Collection restructuring and exception handling
Significant changes to `BindingList.cs`, `BindingSet.cs`, `CiccioList.cs`, `CiccioSet.cs`, and others to improve event handling and the binding interface. Removal and modification of methods and properties to simplify code and improve compatibility with `IBindingList` and `IRaiseItemChangedEvents`.

Updated documentation in `README.txt` to reflect changes and location of built packages.